Added is_standalone? guard to drdb and added --smoke for default-org pedant test
drdb recipe had been getting included because standalone counts as
backend. This adds a is_standalone? guard to ensure that it does not get
kicked off.

Added --smoke function for default-org. The understanding is that in the
future, there will be some sort of flag indicating whether pedant should
be run in full mode or smoke mode. Since the default is smoke mode,
default-org pedant tests should be run in smoke mode.